refactor: WidgetEndpoint.php still used deprecated Widget.php
Created new HasWidget.php trait to make functionality available for the WidgetController.php and the WidgetEndpoint.php. Small change was needed for WidgetScriptBuilder.php as the endpoint had to return the widget without the HTML wrapper.

Use method withHTML() on the builder to retrieve the widget with the script html wrapper.