fix: checkboxes now show correct state
For this fix we needed serverside AND clientside fixes. Serverside: changed from empty() to !isset check because a zero or false value is still a value. Clientside: we do no longer transform data in the hook to make checkbox values always a boolean, we just made the checkbox component work with truthy or falsy values.