fix: authenticate from the top when "refresh token" request is unauthorized
That can be the case when our refresh token is invalid, maybe a new one was requested or ours is outdated. If even this request turns out invalid a user probably changed their password. Ifso, we set a flag to true in the database and prevent executing calls. To be able to do this we have to save their random password in the company data, we do this encrypted.