fix: Calendar published task and notice were always shown after onboarding
This was because the "Calendar published" event was triggered before Tasks and Notices were constructed. And thus the Event we called had no effect on these tasks, defaulting them to active and urgent. Now using a flag in the options table that can be read during construct of the task and notice. Based on the flag we detemrine the default status.