fix: prevent silent padding of key + make sure key is 32 bytes (fixes logout issues)
Previously our key was not compliant with the algorithm we used. That required a key of 32 bytes and if it was shorter the openssl_encrypt method would pad the key silently. Which could result in unexpected data which resulted in empty decrypted responses. Which in turns "logs-out" a user.

Sources:
- https://www.php.net/manual/en/openssl.constants.other.php#constant.openssl-zero-padding
- https://www.cryptosys.net/manapi/api_blockciphersizes.html